Patched for Swift 1.2. (Xcode 6.3b2)

+++ Project/Subprojects/LLDBWrapper/LLDBWrapper/LLDBExtensions.swift
@@ -544,7 +544,7 @@ public extension LLDBData {
 			d.length	=	Int(self.byteSize)
 			let	sz1		=	self.readRawDataWithOffset(LLDBOffsetType(0), buffer: d.mutableBytes, size: self.byteSize, error: &e)
 			precondition(e == nil, "An error `\(e)` occured while reading from this object.")
-			precondition(sz1 == UInt(d.length))
+			precondition(sz1 == d.length)
 			precondition(sz1 == self.byteSize)
 			return	d
 		}
